separated server.js
relates to #25
