+++ src/routes/home.js
@@ -1,0 +1,49 @@
+
+var express = require('express');
+var router = express.Router();
+
+  router.get('/', (req, res) => {
+  res.send('I should be working');
+});
+
+router.get('/webhook', (req, res) => {
+  if (req.query['hub.mode'] === 'subscribe' &&
+      req.query['hub.verify_token'] === process.env.FB_SECRET) {
+    console.log('Validating webhook');
+    res.status(200).send(req.query['hub.challenge']);
+  } else {
+    console.error('Failed validation. Make sure the validation tokens match.');
+    res.sendStatus(403);
+  }
+});
+
+router.post('/webhook', (req, res) => {
+  const data = req.body;
+
+  // Make sure this is a page subscription
+  if (data.object === 'page') {
+    // Iterate over each entry - there may be multiple if batched
+    data.entry.forEach((entry) => {
+      const pageID = entry.id;
+      const timeOfEvent = entry.time;
+
+      // Iterate over each messaging event
+      entry.messaging.forEach((event) => {
+        if (event.message) {
+          receivedMessage(event);
+        } else {
+          console.log('Webhook received unknown event: ', event);
+        }
+      });
+    });
+
+    // Assume all went well.
+    //
+    // You must send back a 200, within 20 seconds, to let us know
+    // you've successfully received the callback. Otherwise, the request
+    // will time out and we will keep trying to resend.
+    res.sendStatus(200);
+  }
+});
+
+module.exports = router;
