+++ webhook.js
@@ -6,62 +6,11 @@ require('env2')('./config.env');
 const apiai = require('apiai');
 const apiai_app = apiai(process.env.APIAI_CLIENT);
 
+
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
-const server = app.listen(process.env.PORT || 4000, () => {
-  console.log('Express server listening on port %d in %s mode', server.address().port, app.settings.env);
-});
 
-/* Home */
-app.get('/', (req, res) => {
-  res.send('I should be working');
-  res.status(200);
-});
-
-/* For Facebook Validation */
-app.get('/webhook', (req, res) => {
-  if (req.query['hub.mode'] === 'subscribe' &&
-      req.query['hub.verify_token'] === process.env.FB_SECRET) {
-    console.log('Validating webhook');
-    res.status(200).send(req.query['hub.challenge']);
-  } else {
-    console.error('Failed validation. Make sure the validation tokens match.');
-    res.sendStatus(403);
-  }
-});
-
-
-// Listen for messages from user
-
-app.post('/webhook', (req, res) => {
-  const data = req.body;
-
-  // Make sure this is a page subscription
-  if (data.object === 'page') {
-    // Iterate over each entry - there may be multiple if batched
-    data.entry.forEach((entry) => {
-      const pageID = entry.id;
-      const timeOfEvent = entry.time;
-
-      // Iterate over each messaging event
-      entry.messaging.forEach((event) => {
-        if (event.message) {
-          receivedMessage(event);
-        } else {
-          console.log('Webhook received unknown event: ', event);
-        }
-      });
-    });
-
-    // Assume all went well.
-    //
-    // You must send back a 200, within 20 seconds, to let us know
-    // you've successfully received the callback. Otherwise, the request
-    // will time out and we will keep trying to resend.
-    res.sendStatus(200);
-  }
-});
 
 function receivedMessage(event) {
   const senderID = event.sender.id;
@@ -132,5 +81,3 @@ function callSendAPI(messageData) {
     }
   });
 }
-
-module.exports = server;
